AT [Problem] max integer smaller than values provided by test 100000 x 90000

+++ max_pairwise_product/main.cpp
@@ -5,13 +5,13 @@ using std::vector;
 using std::cin;
 using std::cout;
 
-long long MaxPairwiseProduct(const vector<int>& numbers) {
-    long long result = 0;
+int MaxPairwiseProduct(const vector<int>& numbers) {
+    int result = 0;
     int n = numbers.size();
     for (int i = 0; i < n; ++i) {
         for (int j = i + 1; j < n; ++j) {
-            if (((long long)numbers[i]) * numbers[j] > result) {
-                result = ((long long)numbers[i]) * numbers[j];
+            if (numbers[i] * numbers[j] > result) {
+                result = numbers[i] * numbers[j];
             }
         }
     }
@@ -26,7 +26,7 @@ int main() {
         cin >> numbers[i];
     }
     
-    long long result = MaxPairwiseProduct(numbers);
+    int result = MaxPairwiseProduct(numbers);
     cout << result << "\n";
     return 0;
 }
